// packages/shared/src/enums/game.enums.d.ts
export declare enum GameId {
    Quiz = "quiz",
    FindDifference = "find-difference",
    RockPaperScissors = "rock-paper-scissors",
    PriceCompare = "price-compare",
    TriviaConquest = "trivia-conquest"
}
export declare enum GameStatus {
    Idle = "idle",
    Playing = "playing",
    Paused = "paused",
    Finished = "finished"
}
export declare enum RoundOutcome {
    Win = "win",
    Lose = "lose",
    Draw = "draw"
}
//# sourceMappingURL=game.enums.d.ts.map